runtbot/lbomon.py
#!/usr/bin/env python3

import RPi.GPIO as GPIO
from time import sleep

GPIO.setmode(GPIO.BCM)
GPIO.setup(18, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)

while not GPIO.input(18):
    sleep(10)

print("Detected low battery power.")
