requirements.txt
RPi.GPIO
flask
requests
click
